@@ -41,21 +41,21 @@
     "@types/node": "^12.11.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.901.0-next.3",
-    "@angular-devkit/build-angular": "^0.901.0-next.3",
-    "@angular-devkit/core": "^9.1.0-next.3",
-    "@angular-devkit/schematics": "^9.1.0-next.3",
-    "@angular/animations": "^9.1.0-next.4",
-    "@angular/bazel": "^9.1.0-next.4",
-    "@angular/cli": "^9.1.0-next.3",
-    "@angular/common": "^9.1.0-next.4",
-    "@angular/compiler": "^9.1.0-next.4",
-    "@angular/compiler-cli": "^9.1.0-next.4",
-    "@angular/core": "^9.1.0-next.4",
-    "@angular/platform-browser": "^9.1.0-next.4",
-    "@angular/platform-browser-dynamic": "^9.1.0-next.4",
-    "@angular/platform-server": "^9.1.0-next.4",
-    "@angular/router": "^9.1.0-next.4",
+    "@angular-devkit/architect": "^0.1000.0-next.4",
+    "@angular-devkit/build-angular": "^0.1000.0-next.4",
+    "@angular-devkit/core": "^10.0.0-next.4",
+    "@angular-devkit/schematics": "^10.0.0-next.4",
+    "@angular/animations": "^10.0.0-next.4",
+    "@angular/bazel": "^10.0.0-next.4",
+    "@angular/cli": "^10.0.0-next.4",
+    "@angular/common": "^10.0.0-next.4",
+    "@angular/compiler": "^10.0.0-next.4",
+    "@angular/compiler-cli": "^10.0.0-next.4",
+    "@angular/core": "^10.0.0-next.4",
+    "@angular/platform-browser": "^10.0.0-next.4",
+    "@angular/platform-browser-dynamic": "^10.0.0-next.4",
+    "@angular/platform-server": "^10.0.0-next.4",
+    "@angular/router": "^10.0.0-next.4",
     "@bazel/bazel": "2.1.0",
     "@bazel/buildifier": "^3.0.0",
     "@bazel/hide-bazel-files": "~1.6.0",
@@ -64,7 +64,7 @@
     "@bazel/karma": "~1.6.0",
     "@bazel/typescript": "~1.6.0",
     "@hapi/hapi": "^18.4.0",
-    "@schematics/angular": "^9.1.0-next.3",
+    "@schematics/angular": "^10.0.0-next.4",
     "@types/browser-sync": "^2.26.1",
     "@types/express": "4.17.6",
     "@types/express-serve-static-core": "4.17.7",
